--- MvvX.Plugins.AppCenter.Platform/MvvX.Plugins.AppCenter.Platform.csproj
@@ -1,16 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <TargetFramework>netstandard2.0</TargetFramework>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <PackageReference Include="Microsoft.AppCenter.Analytics" Version="2.3.0-preview" />
-    <PackageReference Include="Microsoft.AppCenter.Crashes" Version="2.3.0-preview" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\MvvX.Plugins.AppCenter\MvvX.Plugins.AppCenter.csproj" />
-  </ItemGroup>
-
-</Project>
