--- nuspec/BootstrapContent/AppCenterPluginBootstrap.cs.pp
@@ -1,9 +1,0 @@
-using MvvmCross.Platform.Plugins;
-
- namespace $rootnamespace$.Bootstrap
- {
-    public class AppCenterPluginBootstrap
-        : MvxPluginBootstrapAction<MvvX.Plugins.AppCenter.PluginLoader>
-    {
-    }
- }
