--- nuspec/BootstrapContentTouch/AppCenterPluginBootstrap.cs.pp
@@ -1,9 +1,0 @@
-using MvvmCross.Platform.Plugins;
-
- namespace $rootnamespace$.Bootstrap
- {
-    public class AppCenterPluginBootstrap
-        : MvxLoaderPluginBootstrapAction<MvvX.Plugins.AppCenter.PluginLoader, MvvX.Plugins.AppCenter.Touch.Plugin>
-    {
-    }
- }
