--- nuspec/MvvX.Plugins.AppCenter.nuspec
@@ -1,59 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
- <package xmlns="http://schemas.microsoft.com/packaging/2010/07/nuspec.xsd">
-    <metadata>
-        <id>MvvX.Plugins.AppCenter</id>
-        <version>1.0.0.0</version>
-        <title>AppCenter mvvmcross plugin</title>
-        <authors>MACK Mathieu</authors>
-        <owners>MACK Mathieu</owners>
-        <licenseUrl>https://github.com/mathieumack/MvvX.Plugins.AppCenter</licenseUrl>
-        <projectUrl>https://github.com/mathieumack/MvvX.Plugins.AppCenter</projectUrl>
-        <requireLicenseAcceptance>false</requireLicenseAcceptance>
-        <description>This package contains the AppCenter plugin for MvvmCross (pcl)</description>
-        <tags>mvvmcross plugin appcenter</tags>
-        <dependencies>
-            <group>
-                <!-- Always include regardless of target framework -->
-                <dependency id="MvvmCross.Platform" version="4.2.2" />
-            </group>
-            <group targetFramework="net45">
-                <dependency id="Microsoft.AppCenter.Analytics" version="2.3.0-preview" />
-                <dependency id="Microsoft.AppCenter.Crashes" version="2.3.0-preview" />
-            </group>
-            <group targetFramework="MonoAndroid">
-              <dependency id="Microsoft.AppCenter.Analytics" version="2.1.1" />
-              <dependency id="Microsoft.AppCenter.Crashes" version="2.1.1" />
-            </group>
-            <group targetFramework="MonoTouch10">
-              <dependency id="Microsoft.AppCenter.Analytics" version="2.1.1" />
-              <dependency id="Microsoft.AppCenter.Crashes" version="2.1.1" />
-            </group>
-        </dependencies>  
-    </metadata>
-    <files>     
-        <!-- ALL -->
-        <file src="..\MvvX.Plugins.AppCenter\bin\Release\MvvX.Plugins.AppCenter.XML" target="lib\portable-net45+sl5+wp8+wpa81+win8+monoandroid+monotouch" />
-        <file src="..\MvvX.Plugins.AppCenter\bin\Release\MvvX.Plugins.AppCenter.dll" target="lib\portable-net45+sl5+wp8+wpa81+win8+monoandroid+monotouch" />
-
-        <!-- Wpf -->
-        <file src="..\MvvX.Plugins.AppCenter\bin\Release\MvvX.Plugins.AppCenter.XML" target="lib\net45" />
-        <file src="..\MvvX.Plugins.AppCenter\bin\Release\MvvX.Plugins.AppCenter.dll" target="lib\net45" />
-        <file src="..\MvvX.Plugins.AppCenter.Platform\bin\Release\netstandard2.0\MvvX.Plugins.AppCenter.Platform.dll" target="lib\net45" />
-        <file src="..\MvvX.Plugins.AppCenter.Wpf\bin\Release\MvvX.Plugins.AppCenter.Wpf.dll" target="lib\net45" />
-        <file src="BootstrapContent\AppCenterPluginBootstrap.cs.pp" target="content\net45\Bootstrap\AppCenterPluginBootstrap.cs.pp" />
-        
-        <!-- Android -->
-        <file src="..\MvvX.Plugins.AppCenter\bin\Release\MvvX.Plugins.AppCenter.XML" target="lib\MonoAndroid" />
-        <file src="..\MvvX.Plugins.AppCenter\bin\Release\MvvX.Plugins.AppCenter.dll" target="lib\MonoAndroid" />
-        <file src="..\MvvX.Plugins.AppCenter.Platform\bin\Release\netstandard2.0\MvvX.Plugins.AppCenter.Platform.dll" target="MonoAndroid" />
-        <file src="..\MvvX.Plugins.AppCenter.Droid\bin\Release\MvvX.Plugins.AppCenter.Droid.dll" target="lib\MonoAndroid" />
-        <file src="BootstrapContent\AppCenterPluginBootstrap.cs.pp" target="content\MonoAndroid\Bootstrap\AppCenterPluginBootstrap.cs.pp" />
-        
-        <!-- Touch -->
-        <file src="..\MvvX.Plugins.AppCenter\bin\Release\MvvX.Plugins.AppCenter.XML" target="lib\Xamarin.iOS10" />
-        <file src="..\MvvX.Plugins.AppCenter\bin\Release\MvvX.Plugins.AppCenter.dll" target="lib\Xamarin.iOS10" />
-        <file src="..\MvvX.Plugins.AppCenter.Platform\bin\Release\netstandard2.0\MvvX.Plugins.AppCenter.Platform.dll" target="lib\Xamarin.iOS10" />
-        <file src="..\MvvX.Plugins.AppCenter.Touch\bin\Release\MvvX.Plugins.AppCenter.Touch.dll" target="lib\Xamarin.iOS10" />
-        <file src="BootstrapContentTouch\AppCenterPluginBootstrap.cs.pp" target="content\Xamarin.iOS10\Bootstrap\AppCenterPluginBootstrap.cs.pp" />
-    </files>
-</package>
